Organize components into directories
modified:   src/App.js
renamed:    src/Components/CategoryCard.css -> src/Components/CategoryCard/CategoryCard.css
renamed:    src/Components/CategoryCard.js -> src/Components/CategoryCard/CategoryCard.js
renamed:    src/Components/CategoryList.css -> src/Components/CategoryList/CategoryList.css
renamed:    src/Components/CategoryList.js -> src/Components/CategoryList/CategoryList.js
renamed:    src/Components/Fetching.css -> src/Components/Fetching/Fetching.css
renamed:    src/Components/Fetching.js -> src/Components/Fetching/Fetching.js
renamed:    src/Components/Header.css -> src/Components/Header/Header.css
renamed:    src/Components/Header.js -> src/Components/Header/Header.js
renamed:    src/Components/Leader.css -> src/Components/Leader/Leader.css
renamed:    src/Components/Leader.js -> src/Components/Leader/Leader.js
renamed:    src/Components/LeaderBoard.css -> src/Components/LeaderBoard/LeaderBoard.css
renamed:    src/Components/LeaderBoard.js -> src/Components/LeaderBoard/LeaderBoard.js
renamed:    src/Components/PlayerCard.css -> src/Components/PlayerCard/PlayerCard.css
renamed:    src/Components/PlayerCard.js -> src/Components/PlayerCard/PlayerCard.js
renamed:    src/Components/ProgressCircle.css -> src/Components/ProgressCircle/ProgressCircle.css
renamed:    src/Components/ProgressCircle.js -> src/Components/ProgressCircle/ProgressCircle.js
renamed:    src/Components/SearchBar.css -> src/Components/SearchBar/SearchBar.css
renamed:    src/Components/SearchBar.js -> src/Components/SearchBar/SearchBar.js
modified:   src/Pages/HomePage.js
